pull upload-specific code into its own module
start using exceptions for better flow control and separating the
processing logic from the error rendering logic

This commit is 90% just moving code around.
